Update and rename IntakeCommand.java to IntakeOnCommand.java
